dlog: add write delay failure injection options - per stream or just percent selection - use very large delay on stream to simulate high pending - smaller delays for progress with high latency
RB_ID=561924

+++ src/main/java/com/twitter/distributedlog/Utils.java
@@ -66,8 +66,8 @@ public class Utils {
         return (System.currentTimeMillis() - startMsecTime);
     }
 
-    public static boolean randomPercent(int percent) {
-        return (Math.random() * 100) <= percent;
+    public static boolean randomPercent(double percent) {
+        return (Math.random() * 100.0) <= percent;
     }
 
     public static void zkCreateFullPathOptimistic(
